requirements.txt
# HTTP requests
requests>=2.31.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0

# Database
sqlalchemy>=2.0.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Retry logic
tenacity>=8.2.0

# Excel export
openpyxl>=3.1.0

# Text processing (lemmatization)
pymorphy3>=1.0.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0

# Logging enhancement
colorlog>=6.7.0

# Web API (FastAPI)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
aiofiles>=23.2.1

# Background tasks
anyio>=4.0.0

# Progress bar
tqdm>=4.66.0
